requirements.txt
discord.py
httpx
SQLAlchemy
asyncpg
python-dotenv
aiohttp
fastapi
uvicorn
jinja2
python-multipart
requests-oauthlib
authlib
itsdangerous
apscheduler
psycopg2
pymysql